requirements.txt
google-adk>=0.0.1
google-cloud-aiplatform[adk,agent-engines]>=1.88.0
google-cloud-storage>=2.8.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
vertexai>=0.0.1
requests>=2.31.0
python-dotenv
beautifulsoup4
sqlalchemy>=2.0
pymysql>=1.1.0
google-genai "google-cloud-aiplatform[evaluation]"
google-adk[eval]
# google-cloud-vision
# Pillow
